Moved serializer from forms to make_a_plea
