+++ apps/forms/tests.py
@@ -1,5 +1,4 @@
-from datetime import datetime, date
-from mock import patch, MagicMock
+from mock import patch
 
 from django import forms
 from django.contrib import messages
@@ -7,7 +6,6 @@ from django.forms.formsets import formset_factory
 from django.http import Http404
 from django.test import TestCase
 from .stages import MultiStageForm, FormStage
-from .serializers import DateAwareSerializer
 
 
 def reverse(url_name, args=None):
@@ -223,29 +221,3 @@ class TestMultiStageForm(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "Enter a whole number")
         self.assertContains(response, "This field is required")
-
-
-class DateAwareSerializerTests(TestCase):
-    def test_serializer_dumps_plain(self):
-        test_str = """{"key2": 2, "key1": "value1"}"""
-        test_dict = {"key1": "value1",
-                     "key2": 2}
-
-        ds = DateAwareSerializer().dumps(test_dict)
-        self.assertEqual(ds, test_str)
-
-    def test_serializer_loads_plain(self):
-        test_str = """{"key2": 2, "key1": "value1"}"""
-        test_dict = {"key1": "value1",
-                     "key2": 2}
-
-        ds = DateAwareSerializer().loads(test_str)
-        self.assertEqual(ds, test_dict)
-
-    def test_serializer_dumps_datetime(self):
-        test_str = """{"key2": "2014-06-03T00:00:00", "key1": "value1"}"""
-        test_dict = {"key1": "value1",
-                     "key2": datetime(2014, 6, 3, 0, 0)}
-
-        ds = DateAwareSerializer().dumps(test_dict)
-        self.assertEqual(ds, test_str)
