+++ make_a_plea/settings/base.py
@@ -145,7 +145,7 @@ MIDDLEWARE_CLASSES = (
 
 ROOT_URLCONF = 'make_a_plea.urls'
 
-SESSION_SERIALIZER = 'apps.forms.serializers.DateAwareSerializer'
+SESSION_SERIALIZER = 'make_a_plea.serializers.DateAwareSerializer'
 SESSION_ENGINE = 'django.contrib.sessions.backends.signed_cookies'
 SESSION_COOKIE_HTTPONLY = True
 SESSION_EXPIRE_AT_BROWSER_CLOSE = True
